Run activation script in syschdemd; don't use custom wrapper dir
Otherwise the activation script fails in the "specialfs" phase.
This happens because NixOS mounts wrapperDir's parent directory as
tmpfs. With "/wrappers" as value, that means "/" would be tried to be
(re-)mounted as tmpfs, and that has little chances of success.

Luckily, a custom value for `wrapperDir` isn't even necessary and was
only done because I grossly misunderstood NixOS's boot process: It calls
the activation script on each boot. As syschdemd is NixOS-WSL's init
replacement, calling the activation script there should be the correct
fix.

Fixes #11