Fix symlink creation of per-user apps/icons folder
The code relied on `defaultUser` attribute name in `config.users.users`
to be the same as the username. This is not always the case as the
username can be set by `name` attribute inside the user configuration

So for following user

```
users.users.foo.name = "bar"
```

the `wsl.defaultUser = "foo"` works, but the `wsl.startMenuLaunchers =
true` won't work because it tries to find applications folder at
`/etc/profiles/per-user/foo/share` rather than
`/etc/profiles/per-user/bar/share` that is actually created.

This uses the `name` attribute of that user instead.