docker-desktop: do not install docker{,-compose}
Docker Desktop comes with its own 'docker' and CLI plugin executables
already, including Compose. Docker Desktop symlinks those executables in
the /usr/bin directory of the WSL instance whenever it is up and removes
them when it shuts down.

If the user wants to install their own 'docker' program they are free to
do so, but it is by no means something that Docker Desktop requires.